% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyph.support-es.R
\name{hyph.support.es}
\alias{hyph.support.es}
\title{Language support for Spanish}
\usage{
hyph.support.es()
}
\description{
This function adds support for Spanish to the sylly package. You should not
need to call it manually, as that is done automatically when this package is
loaded.
}
\details{
In particular,
      a new set of hyphenation patterns is being added (see \code{\link{hyph.es}}).
To use the patterns with \code{\link[sylly:hyphen]{hyphen}}, simply use the abbreviation:
\itemize{
  \item {\code{"es"}} {--- Spanish hyphenation patterns}
}
}
\seealso{
\code{\link[sylly:hyphen]{hyphen}},
   \code{\link[sylly:set.hyph.support]{set.hyph.support}}
}
